+++ Changelog.md
@@ -13,7 +13,7 @@
 
 ### Enhancements
 
-* Added `yieldRate` combinator to control the yield rate of a stream.
+* Added rate limiting combinator `rate` to control the yield rate of a stream.
 * Add `foldl1'`, `foldr1`, `intersperseM`, `find`, `lookup`, `and`, `or`,
   `findIndices`, `findIndex`, `elemIndices`, `elemIndex`, `init` to Prelude